# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  dtos.MailStatusDTO:
    properties:
      duration_seconds:
        type: number
      error:
        example: smtp auth failed
        type: string
      expires_at:
        type: string
      expires_in_seconds:
        type: integer
      result:
        example: Email sent to user@example.com
        type: string
      started_at:
        type: string
      status:
        example: completed
        type: string
      trigger:
        type: string
    type: object
...
  dtos.RAGRawPromptRequestDTO:
    properties:
      prompt:
        example: Hello, how are you?
        type: string
    required:
    - prompt
    type: object
  dtos.RAGRawPromptResponseDTO:
    properties:
      duration_seconds:
        type: number
      error:
        type: string
      result:
        type: string
      started_at:
        type: string
      status:
        example: completed
        type: string
      trigger:
        type: string
    type: object
  dtos.RAGRequestDTO:
    properties:
      language:
        type: string
      text:
        type: string
    required:
    - text
    type: object
  dtos.RAGStatusDTO:
    properties:
      body: {}
      duration_seconds:
        example: 2.5
        type: number
      endpoint:
        type: string
      error:
        example: gemini api returned status 503
        type: string
      execution_result:
        description: holds the response from the fetched endpoint
      expires_at:
        type: string
      expires_in_seconds:
        type: integer
      headers:
        additionalProperties:
          type: string
        type: object
      method:
        type: string
      result:
        example: The meeting discussed...
        type: string
      started_at:
        example: "2026-02-21T11:00:00Z"
        type: string
      status:
        example: completed
        type: string
      trigger:
        example: /api/rag/summary
        type: string
    type: object
  dtos.RAGSummaryRequestDTO:
    properties:
      context:
        example: technical meeting
        type: string
      date:
        type: string
      language:
        description: '"id" or "en"'
        example: id
        type: string
      location:
        example: Meeting Room A
        type: string
      mac_address:
        example: AA:BB:CC:DD:EE:FF
        type: string
      participants:
        example:
        - '["Alice"'
        - ' "Bob"]'
        items:
          type: string
        type: array
      style:
        description: e.g., "minutes", "executive"
        example: minutes
        type: string
      text:
        example: This is a long transcript of a technical meeting...
        type: string
    required:
    - text
    type: object
...
  recordings_dtos.GetAllRecordingsResponseDto:
    properties:
      limit:
        type: integer
      page:
        type: integer
      recordings:
        items:
          $ref: '#/definitions/recordings_dtos.RecordingResponseDto'
        type: array
      total:
        type: integer
    type: object
  recordings_dtos.RecordingResponseDto:
    properties:
      audio_url:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      original_name:
        type: string
    type: object
  recordings_dtos.StandardResponse:
    properties:
      data: {}
      details:
        description: |-
          Details is only populated for 400 (Bad Request) and 422 (Unprocessable Entity) errors.
          For all other status codes, including 500, this field is nil/omitted.
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
  sensio_domain_common_dtos.StandardResponse:
    properties:
      data: {}
      details:
        description: |-
          Details is only populated for 400 (Bad Request) and 422 (Unprocessable Entity) errors.
          For all other status codes, including 500, this field is nil/omitted.
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
  sensio_domain_rag_dtos.StandardResponse:
    properties:
      data: {}
      details: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Sensio App. <br> For full documentation,
    visit <a href="/docs">/docs</a>.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sensio API
  version: "1.0"
paths:
...
  /api/mail/status/{task_id}:
    get:
      description: Get the status and result of an email sending task.
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.MailStatusDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get email task status
      tags:
      - 08. Mail
  /api/models/gemini:
    post:
      consumes:
      - application/json
      description: Send a raw prompt directly to the Gemini LLM model without RAG
        orchestration.
      parameters:
      - description: Prompt Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RAGRawPromptRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.RAGRawPromptResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Raw prompt query to Gemini model
      tags:
      - 06. Models
...
  /api/rag/{task_id}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.RAGStatusDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get RAG task status
      tags:
      - 05. RAG
...
  /api/rag/summary:
    post:
      consumes:
      - application/json
      description: Generate meeting minutes summary asynchronously. Returns a Task
        ID for polling.
      parameters:
      - description: Summary request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RAGSummaryRequestDTO'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Generate meeting minutes summary
      tags:
      - 05. RAG
  /api/rag/translate:
    post:
      consumes:
      - application/json
      description: Translate text to a target language asynchronously. Returns a Task
        ID for polling.
      parameters:
      - description: Translation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RAGRequestDTO'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_rag_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Translate text to specified language
      tags:
      - 05. RAG
  /api/recordings:
    get:
      description: Get a paginated list of all recordings.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/recordings_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/recordings_dtos.GetAllRecordingsResponseDto'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all recordings
      tags:
      - 07. Recordings
    post:
      consumes:
      - multipart/form-data
      description: Upload an audio file and save its metadata.
      parameters:
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/recordings_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/recordings_dtos.RecordingResponseDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Save a new recording
      tags:
      - 07. Recordings
  /api/recordings/{id}:
    delete:
      description: Remove a recording and its associated file from the system.
      parameters:
      - description: Recording ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a recording
      tags:
      - 07. Recordings
    get:
      description: Retrieve metadata for a specific recording.
      parameters:
      - description: Recording ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/recordings_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/recordings_dtos.RecordingResponseDto'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get recording by ID
      tags:
      - 07. Recordings
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...