# backend/domain/rag/controllers/rag_status_controller.go
package controllers

import (
	"net/http"
	"sensio/domain/common/tasks"
	"sensio/domain/rag/dtos"

	"github.com/gin-gonic/gin"
)

// RAGStatusController handles task status requests.
type RAGStatusController struct {
	statusUC tasks.GenericStatusUseCase[dtos.RAGStatusDTO]
}

func NewRAGStatusController(statusUC tasks.GenericStatusUseCase[dtos.RAGStatusDTO]) *RAGStatusController {
	return &RAGStatusController{
		statusUC: statusUC,
	}
}

// GetStatus handles GET /api/rag/:task_id
// @Summary Get RAG task status
// @Tags 05. RAG
// @Security BearerAuth
// @Produce json
// @Param task_id path string true "Task ID"
// @Success 200 {object} dtos.StandardResponse{data=dtos.RAGStatusDTO}
// @Failure 404 {object} dtos.StandardResponse
// @Router /api/rag/{task_id} [get]
func (c *RAGStatusController) GetStatus(ctx *gin.Context) {
	id := ctx.Param("task_id")
	status, err := c.statusUC.GetTaskStatus(id)
	if err != nil {
		ctx.JSON(http.StatusNotFound, dtos.StandardResponse{
			Status:  false,
			Message: "Resource Not Found",
		})
		return
	}

	isSuccess := status.Status != "failed"
	message := "Task status retrieved successfully"
	httpStatus := http.StatusOK

	if status.Status == "failed" {
		message = "Task failed"
		if status.HTTPStatusCode != 0 {
			httpStatus = status.HTTPStatusCode
		}
	}

	ctx.JSON(httpStatus, dtos.StandardResponse{
		Status:  isSuccess,
		Message: message,
		Data:    status,
	})
}